Feat: separate secondary structure processing into DictMethodSecStructUtils.py.  Pass the dictionary API instance from DictMethodRunner() as a kwarg to all helper classes.  Don't sort NeighborInteractionProvider sucesses.
