+++ rcsb/utils/dictionary/DictMethodEntityHelper.py
@@ -49,8 +49,9 @@ class DictMethodEntityHelper(object):
         #
         rP = kwargs.get("resourceProvider")
         self.__commonU = rP.getResource("DictMethodCommonUtils instance") if rP else None
-        dapw = rP.getResource("DictionaryAPIProviderWrapper instance") if rP else None
-        self.__dApi = dapw.getApiByName("pdbx_core") if dapw else None
+        # dapw = rP.getResource("DictionaryAPIProviderWrapper instance") if rP else None
+        # self.__dApi = dapw.getApiByName("pdbx_core") if dapw else None
+        self.__dApi = kwargs.get("dictionaryApi", None)
         #
         self.__useSiftsAlign = rP.getReferenceSequenceAlignmentOpt() == "SIFTS"
         # logger.info("SIFTS alignment option %r", self.__useSiftsAlign)
