+++ requirements.txt
@@ -1,6 +1,6 @@
 scipy
 numpy
-mmcif >= 0.71
+mmcif >= 0.75
 rcsb.utils.chemref >= 0.79
 rcsb.utils.citation >= 0.16
 rcsb.utils.config >= 0.35
